Track dependencies for streamed transactions
This commit allows tracking dependencies of streamed transactions.

Regarding the streaming=on case, dependency tracking is enabled while applying
spooled changes from files.

In the streaming=parallel case, dependency tracking is performed when the leader
sends changes to parallel workers, just as in non-streaming transactions. Apart
from non-streamed transactions, the leader waits for parallel workers till the
assigned transactions are finished at COMMIT/PREPARE/ABORT; thus, the XID of
streamed transactions is not cached as the lastly handled one.

XXX: XID of top-transaction is always used for dependency tracking. This means
another parallel worker waits for a dependent transaction, even if it is a
sub-transaction, and it is actually rolled back by "ROLLBACK TO SAVEPOINT".